Add some initial (probably wrong) QC
This patch hacks in a pass to `grep -v` out the overrepresented
sequences reported by FastQC.  However, it does it on a per-FASTQ basis,
which probably isn't right: since these are paired-end reads, if one of
them is removed we also need to remove its mate from the other file.

We should fix this at some point, but none of the FastX toolkit tools
seem to handle this, so we'll need to look elsewhere.